refactor assembler into per-expression modules
