+++ src/assembler/compile-identifier.ts
@@ -1,0 +1,26 @@
+import { CompileExprOpts, mapBinaryenType } from "../assembler.js";
+import { Identifier } from "../syntax-objects/identifier.js";
+import { refCast } from "../lib/binaryen-gc/index.js";
+
+export const compile = (opts: CompileExprOpts<Identifier>) => {
+  const { expr, mod } = opts;
+
+  if (expr.is("break")) return mod.br(opts.loopBreakId!);
+
+  const entity = expr.resolve();
+  if (!entity) {
+    throw new Error(`Unrecognized symbol ${expr.value}`);
+  }
+
+  if (entity.isVariable() || entity.isParameter()) {
+    const type = mapBinaryenType(opts, entity.originalType ?? entity.type!);
+    const get = mod.local.get(entity.getIndex(), type);
+    if (entity.requiresCast) {
+      return refCast(mod, get, mapBinaryenType(opts, entity.type!));
+    }
+    return get;
+  }
+
+  throw new Error(`Cannot compile identifier ${expr}`);
+};
+
