+++ src/assembler/compile-variable.ts
@@ -1,0 +1,13 @@
+import { CompileExprOpts, compileExpression } from "../assembler.js";
+import { Variable } from "../syntax-objects/variable.js";
+
+export const compile = (opts: CompileExprOpts<Variable>) => {
+  const { expr, mod } = opts;
+  return mod.local.set(
+    expr.getIndex(),
+    expr.initializer
+      ? compileExpression({ ...opts, expr: expr.initializer })
+      : mod.nop()
+  );
+};
+
